Begin refactor, transcoding working
This commit includes the beginning stages of a full refactor for the new
Redacted API. Transcoding and creating torrent files work as they did
before, but only for single torrents specified directly by URL.

Support for scanning a user's snatchlist and comparing it to the
download directory is not working because as of yet there is no endpoint
for getting the format/encoding information about the torrents snatched
by a user. In all likelihood a separate request will have to be made to
the API for each snatched torrent to determine what formats are
uploaded, which will make the process very slow for large snatchlists
due to rate limiting (5 requests per 10 seconds).

This refactor is largely incomplete and is pretty messy. There is a lot
of unused/dead code.