Use temporary directory for new torrent files
This solves the problem encountered when torrent files that are
generated and put into the specified torrents directory are
automatically ingested by a torrent client, and the file deleted. This
means the torrent client gets the torrent, but when we go to upload at
the end, the torrent file is gone.

Instead we generate to a temporary file, and then separately copy the
temporary torrent to the desired target directory, and then the script
is free to do what it wants with the temporary one.